+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fix: Fix deprecated `registrar.messenger` call in `SentryFlutterWeb` (#364)
 * Fix: Enable breadcrumb recording mechanism based on platform
 * Remove deprecated member `enableLifecycleBreadcrumbs`. Use `enableAppLifecycleBreadcrumbs` instead.
+* Fix: Enable breadcrumb recording mechanism based on platform ((#366))
+* Remove deprecated member `enableLifecycleBreadcrumbs`. Use `enableAppLifecycleBreadcrumbs` instead. ((#366))
 
 # 4.1.0-nullsafety.1